Revert "Synch robustness branch"
