+++ docs/synthesis_v3.tex
@@ -183,7 +183,7 @@ vector.  In general, for agent $i$, $E_i$ is a $1\times K +
 
 \section{Optimization Problem}
 As shorthand, let $\Omega = [\Omega_1, \Omega_2...\Omega_N]$, $X=[X_1,
-X_2...X_N]$, $W=[W_1, W_2...W_n]$, and $A=[A_1, A_2...A_N]$. 
+X_2...X_N]$, $W=[W_1, W_2...W_n]$ and $A=[A_1, A_2...A_N]$. 
 
 Finally, we can write the welfare function as:
 \begin{equation}
